refactor : ciga 관련 타입 파일 작성
