+++ src/app.features/inventory/types/ciga.ts
@@ -14,3 +14,20 @@ export type ChoType =
 	| 'ㅌ'
 	| 'ㅍ'
 	| 'ㅎ';
+export const CHO_BUTTONS: ChoType[] = [
+	'전체',
+	'ㄱ',
+	'ㄴ',
+	'ㄷ',
+	'ㄹ',
+	'ㅁ',
+	'ㅂ',
+	'ㅅ',
+	'ㅇ',
+	'ㅈ',
+	'ㅊ',
+	'ㅋ',
+	'ㅌ',
+	'ㅍ',
+	'ㅎ',
+];
