+++ src/app.features/inventory/utils/getInitialSound.ts
@@ -1,0 +1,34 @@
+const CHO_LIST = [
+	'ㄱ',
+	'ㄲ',
+	'ㄴ',
+	'ㄷ',
+	'ㄸ',
+	'ㄹ',
+	'ㅁ',
+	'ㅂ',
+	'ㅃ',
+	'ㅅ',
+	'ㅆ',
+	'ㅇ',
+	'ㅈ',
+	'ㅉ',
+	'ㅊ',
+	'ㅋ',
+	'ㅌ',
+	'ㅍ',
+	'ㅎ',
+];
+/**
+ * for 담배 초성 추출
+ * @param str 담배 이름
+ * @return 초성 ex ㄱ
+ */
+export const getInitialSound = (str: string): string => {
+	// 유니코드 한글 시작 위치 -> 44032
+	const word = str.charCodeAt(0) - 44032;
+	const jongseong = word % 28;
+	const jungseong = ((word - jongseong) / 28) % 21;
+	const choseong = ((word - jongseong) / 28 - jungseong) / 21;
+	return CHO_LIST[choseong];
+};
